fix app.go
