+++ cmd/sifnoded/cmd/root.go
@@ -1,0 +1,5 @@
+package cmd
+
+func RootCmd() {
+
+}
